models refactored
