+++ deepface/models/demography/Age.py
@@ -1,7 +1,7 @@
 import os
 import gdown
 import numpy as np
-from deepface.basemodels import VGGFace
+from deepface.models.facial_recognition import VGGFace
 from deepface.commons import package_utils, folder_utils
 from deepface.models.Demography import Demography
 from deepface.commons.logger import Logger
